requirements.txt
tweepy
mysql-connector
pyrebase4
opencv-python
urllib3
pillow
numpy